# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crm/auth/login:
    post:
      operationId: crmLogin
      tags: [crm-auth]
      summary: CRM user login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmLoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /crm/me:
    get:
      operationId: crmGetMe
      tags: [crm-auth]
      summary: Get current CRM user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── CRM Leads ──────────────────────────────────────────
  /crm/leads:
    get:
      operationId: crmGetLeads
      tags: [crm-leads]
      summary: List all leads
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Lead list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmLeadsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: crmCreateLead
      tags: [crm-leads]
      summary: Create a new lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmLeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmLead"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /crm/stats:
    get:
      operationId: crmGetStats
      tags: [crm-leads]
      summary: CRM dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmStatsResponse"

  # ── CRM Tasks ──────────────────────────────────────────
  /crm/tasks:
    get:
      operationId: crmGetTasks
      tags: [crm-tasks]
      summary: List tasks
      security:
        - bearerAuth: []
      parameters:
        - name: leadId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmTask"
    post:
      operationId: crmCreateTask
      tags: [crm-tasks]
      summary: Create a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmTaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmTask"
...
  /crm/users:
    get:
      operationId: crmGetUsers
      tags: [crm-users]
      summary: List CRM users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmUser"
    post:
      operationId: crmCreateUser
      tags: [crm-users]
      summary: Create a CRM user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmUserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmUser"
...
  /crm/links:
    get:
      operationId: crmGetLinks
      tags: [crm-links]
      summary: List submission links (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Link list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmSubmissionLink"
    post:
      operationId: crmCreateLink
      tags: [crm-links]
      summary: Create a submission link (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmLinkInput"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmSubmissionLink"
  /crm/links/{id}:
    patch:
      operationId: crmUpdateLink
      tags: [crm-links]
      summary: Toggle link active/inactive
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmLinkInput"
      responses:
        "200":
          description: Link updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmSubmissionLink"
    delete:
      operationId: crmDeleteLink
      tags: [crm-links]
      summary: Delete a submission link
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"

  # ── Public Lead Submission ──────────────────────────
  /crm/public/submit/{token}:
    get:
      operationId: crmGetSubmissionForm
      tags: [crm-public]
      summary: Get public submission form info
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Form info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmSubmissionFormInfo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: crmSubmitPublicLead
      tags: [crm-public]
      summary: Submit a lead via public link
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmPublicLeadInput"
      responses:
        "201":
          description: Lead submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: Link not found or inactive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContactResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── CRM Schemas ────────────────────────────────────
    CrmLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    CrmLoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/CrmUser"
      required: [token, user]
    CrmUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, sales, va]
        status:
          type: string
          enum: [active, inactive]
        campaignId:
          type: integer
          nullable: true
        campaignName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, role, status, createdAt]
    CrmUserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, sales, va]
        status:
          type: string
          enum: [active, inactive]
        campaignId:
          type: integer
          nullable: true
      required: [name, email, role]
    CrmLead:
      type: object
      properties:
        id:
          type: integer
        sellerName:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        leadSource:
          type: string
          nullable: true
        address:
          type: string
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        propertyType:
          type: string
          nullable: true
        beds:
          type: integer
          nullable: true
        baths:
          type: number
          nullable: true
        sqft:
          type: integer
          nullable: true
        yearBuilt:
          type: integer
          nullable: true
        ownerName:
          type: string
          nullable: true
        lastSaleDate:
          type: string
          nullable: true
        lastSalePrice:
          type: number
          nullable: true
        condition:
          type: integer
          nullable: true
        currentValue:
          type: number
          nullable: true
        estimatedRepairCost:
          type: number
          nullable: true
        arv:
          type: number
          nullable: true
        mao:
          type: number
          nullable: true
        occupancy:
          type: string
          nullable: true
        isRental:
          type: boolean
        rentalAmount:
          type: number
          nullable: true
        reasonForSelling:
          type: string
          nullable: true
        howSoon:
          type: string
          nullable: true
        askingPrice:
          type: number
          nullable: true
        askingPriceText:
          type: string
          nullable: true
        voiceRecordUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, contacted, qualified, negotiating, under_contract, closed, dead]
        assignedTo:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, sellerName, address, status, createdAt, updatedAt]
...
    CrmLeadInput:
      type: object
      properties:
        sellerName:
          type: string
        phone:
          type: string
        email:
          type: string
        leadSource:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        propertyType:
          type: string
        beds:
          type: integer
        baths:
          type: number
        sqft:
          type: integer
        yearBuilt:
          type: integer
          nullable: true
        ownerName:
          type: string
          nullable: true
        lastSaleDate:
          type: string
          nullable: true
        lastSalePrice:
          type: number
          nullable: true
        condition:
          type: integer
        currentValue:
          type: number
        estimatedRepairCost:
          type: number
        arv:
          type: number
        mao:
          type: number
        status:
          type: string
        assignedTo:
          type: integer
          nullable: true
      required: [sellerName, address]
    CrmLeadsResponse:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/CrmLead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [leads, total, page, limit]
...
    CrmTask:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
          nullable: true
        leadAddress:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, completed]
        createdAt:
          type: string
      required: [id, title, status, createdAt]
    CrmTaskInput:
      type: object
      properties:
        leadId:
          type: integer
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, completed]
      required: [title]
    CrmSubmissionLink:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
        label:
          type: string
          nullable: true
        leadSource:
          type: string
          nullable: true
        active:
          type: boolean
        submissionsCount:
          type: integer
        createdAt:
          type: string
        url:
          type: string
      required: [id, token, active, submissionsCount, createdAt, url]
    CrmLinkInput:
      type: object
      properties:
        label:
          type: string
        leadSource:
          type: string
        active:
          type: boolean
    CrmSubmissionFormInfo:
      type: object
      properties:
        label:
          type: string
          nullable: true
        leadSource:
          type: string
          nullable: true
        active:
          type: boolean
      required: [active]
    CrmPublicLeadInput:
      type: object
      properties:
        sellerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        propertyType:
          type: string
        beds:
          type: integer
        baths:
          type: number
        sqft:
          type: integer
        yearBuilt:
          type: integer
          nullable: true
        ownerName:
          type: string
          nullable: true
        lastSaleDate:
          type: string
          nullable: true
        lastSalePrice:
          type: number
          nullable: true
        condition:
          type: integer
        currentValue:
          type: number
        occupancy:
          type: string
          nullable: true
        reasonForSelling:
          type: string
          nullable: true
        howSoon:
          type: string
          nullable: true
        askingPriceText:
          type: string
          nullable: true
          description: Raw asking price text (may be "want an offer" or a dollar amount)
        voiceRecordUrl:
          type: string
          nullable: true
          description: Base64 data URL of an optional voice recording
        message:
          type: string
      required: [sellerName, address, phone]
...
    CrmStatsResponse:
      type: object
      properties:
        totalLeads:
          type: integer
        newLeads:
          type: integer
        underContract:
          type: integer
        closed:
          type: integer
        totalTasks:
          type: integer
        pendingTasks:
          type: integer
        leadsByStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
            required: [status, count]
        recentLeads:
          type: array
          items:
            $ref: "#/components/schemas/CrmLead"
      required: [totalLeads, newLeads, underContract, closed, totalTasks, pendingTasks, leadsByStatus, recentLeads]
...